mod: 重新组织模块名
