+++ seal/_seal.py
