+++ seal/context/__init__.py
@@ -1,1 +1,1 @@
-from .context import WebContext
+from ._context import WebContext
