+++ seal/db/mysql/__init__.py
@@ -1,2 +1,2 @@
-from .chained_query import ChainedQuery, ChainedQuery as Query, ChainedQuery as Q
-from .chained_update import ChainedUpdate, ChainedUpdate as Update, ChainedUpdate as U
+from ._chained_query import ChainedQuery, ChainedQuery as Query, ChainedQuery as Q
+from ._chained_update import ChainedUpdate, ChainedUpdate as Update, ChainedUpdate as U
