+++ seal/db/mysql/_mysql_connector.py
