+++ seal/db/sqlite/_chained_update.py
@@ -1,8 +1,8 @@
 import traceback
 
-from .mysql_connector import MysqlConnector
-from .meta import Meta
-from ..base_chained_update import BaseChainedUpdate
+from ._meta import Meta
+from ._sqlite_connector import SqliteConnector
+from .._base_chained_update import BaseChainedUpdate
 from ...context import WebContext
 from ...model import BaseEntity
 from datetime import datetime
@@ -15,8 +15,8 @@ class ChainedUpdate(BaseChainedUpdate):
         return Meta
 
     def __init__(self, clz=None, table: str = None, logic_delete_col: str = None):
-        super().__init__(clz, '%s', table, logic_delete_col)
-        self._conn = MysqlConnector().get_connection()
+        super().__init__(clz, '?', table, logic_delete_col)
+        self._conn = SqliteConnector().get_connection()
 
     def _get_cursor(self):
         return self._conn.cursor()
@@ -25,18 +25,16 @@ class ChainedUpdate(BaseChainedUpdate):
         self.sets['deleted'] = 1
         self.update()
 
-    def insert(self, entity: BaseEntity = None, data: dict = None, duplicated_key_update=False,
-               reuse_conn: bool = False):
+    def insert(self, entity: BaseEntity = None, data: dict = None, duplicated_ignore=False, reuse_conn: bool = False):
         c = self._get_cursor()
         try:
             if entity is not None:
-                sql, args = self.insert_statement(entity=entity, duplicated_key_update=duplicated_key_update)
+                sql, args = self.insert_statement(entity=entity, duplicated_ignore=duplicated_ignore)
             elif data is not None:
-                sql, args = self.insert_statement(data=data, duplicated_key_update=duplicated_key_update)
+                sql, args = self.insert_statement(data=data, duplicated_ignore=duplicated_ignore)
             else:
                 raise ValueError('null data')
-            affected = c.execute(sql, args)
-            logger.debug(f'#### affected: {affected}')
+            c.execute(sql, args)
             self._conn.commit()
         except Exception as e:
             logger.error(f'数据库操作异常: {e}')
@@ -47,14 +45,13 @@ class ChainedUpdate(BaseChainedUpdate):
                 self._conn.close()
 
     def insert_bulk(self, entity_list: list[BaseEntity] = None, data_list: list[dict] = None,
-                    duplicated_key_update: bool = False, reuse_conn: bool = False):
+                    duplicated_ignore: bool = False, reuse_conn: bool = False):
         if entity_list is None and data_list is None:
             raise ValueError('null data')
 
         c = self._get_cursor()
         try:
-            sql = self.insert_bulk_statement(duplicated_key_update=duplicated_key_update)
-            total_affected = 0
+            sql = self.insert_bulk_statement(duplicated_ignore=duplicated_ignore)
             if entity_list is not None:
                 for entity in entity_list:
                     now = datetime.now()
@@ -62,24 +59,17 @@ class ChainedUpdate(BaseChainedUpdate):
                     entity.create_by = WebContext().uid()
                     entity.create_at = now
                     args = [getattr(entity, col) for col in self.columns(exclude=["id"])]
-                    if duplicated_key_update:
-                        args += args
-                    logger.debug(f'#### args: {args}')
-                    affected = c.execute(sql, args)
-                    total_affected += affected
+                    logger.info(f'#### args: {args}')
+                    c.execute(sql, tuple(args))
             elif data_list is not None:
                 for data in data_list:
                     now = datetime.now()
                     data['deleted'] = 0
                     data['create_by'] = WebContext().uid()
                     data['create_at'] = now
-                    args = [data[col] for col in self.columns(exclude=["id"])]
-                    if duplicated_key_update:
-                        args += args
-                    logger.debug(f'#### args: {args}')
-                    affected = c.execute(sql, args)
-                    total_affected += affected
-            logger.debug(f'#### affected: {total_affected}')
+                    args = [data[col] for col in self.clz.columns(exclude=["id"])]
+                    logger.info(f'#### args: {args}')
+                    c.execute(sql, tuple(args))
             self._conn.commit()
         except Exception as e:
             logger.error(f'数据库操作异常: {e}')
@@ -92,9 +82,9 @@ class ChainedUpdate(BaseChainedUpdate):
     def update(self, reuse_conn: bool = False):
         c = self._get_cursor()
         try:
+
             sql, args = self.update_statement()
-            affected = c.execute(sql, args)
-            logger.debug(f'#### affected: {affected}')
+            c.execute(sql, args)
             self._conn.commit()
         except Exception as e:
             logger.error(f'数据库操作异常: {e}')
@@ -108,8 +98,7 @@ class ChainedUpdate(BaseChainedUpdate):
         c = self._get_cursor()
         try:
             sql, args = self.delete_statement()
-            affected = c.execute(sql, args)
-            logger.debug(f'#### affected: {affected}')
+            c.execute(sql, args)
             self._conn.commit()
         except Exception as e:
             logger.error(f'数据库操作异常: {e}')
@@ -128,9 +117,7 @@ class ChainedUpdate(BaseChainedUpdate):
                 sql, args = self.update_by_pk_statement(data=data)
             else:
                 raise ValueError('null data')
-
-            affected = c.execute(sql, args)
-            logger.debug(f'#### affected: {affected}')
+            c.execute(sql, args)
             self._conn.commit()
         except Exception as e:
             logger.error(f'数据库操作异常: {e}')
