+++ seal/db/sqlite/_sqlite_connector.py
