+++ seal/exception/__init__.py
@@ -1,1 +1,1 @@
-from .business_exception import BusinessException
+from ._business_exception import BusinessException
