+++ seal/exception/_business_exception.py
