+++ seal/model/__init__.py
@@ -1,3 +1,3 @@
-from .base_entity import BaseEntity
-from .response import Response
-from .page_result import PageResult
+from ._base_entity import BaseEntity
+from ._response import Response
+from ._page_result import PageResult
