+++ seal/model/_base_entity.py
