+++ seal/model/_page_result.py
