+++ seal/router/__init__.py
@@ -1,2 +1,2 @@
-from .router import app, get, post, put, delete
-from .jwt import generate_token
+from ._router import app, get, post, put, delete
+from ._jwt import generate_token
