+++ seal/wrapper/__init__.py
@@ -1,3 +1,3 @@
-from .entity import entity
-from .singleton import singleton
-from .cache import cache_func
+from ._entity import entity
+from ._singleton import singleton
+from ._cache import cache_func
