+++ seal/wrapper/_entity.py
