+++ seal/wrapper/_singleton.py
